Simulated Data from Buckley (2002)

In Buckley (2002) tissue residue curves for a Meningioma and a Breast Cancer were simulated using the MMID4 model. Note, the model is described in detail by Bassingthwaighte, J.B. et al. (1984) and Kroll, K et al. (1996). This model accounts for flow dispersion and heterogeneity, and includes capillaries modeled as axially distributed blood-tissue exchange units. A plasma concentration-time curve, AKA arterial input function, was simulated as an input to the model using measurements made by Fritz-Hansen et al. (1996).

data("buckley")

Format

Two lists are created (breast and meningioma) that contain the simulated time curves and all associated kinetic parameter values.

Source

See below.

References

Buckley, D.L. (2002) Uncertainty in the Analysis of Tracer Kinetics Using Dynamic Contrast-Enhanced \(T_1\)-weighted MRI, Magnetic Resonance in Medicine 47, 601-606.

Bassingthwaighte, J.B. and Goresky, C.A. (1984) Modelling in the analysis of solute and water exchange in the microvasculature. In: Renkin, E.M., Michel, C.C. and Geiger, S.R., editors. Handbook of physiology. Section 2. The cardiovascular system. Bethesda: American Physiological Society. p549-626.

Kroll, K., Wilke, N., Jerosch-Herold, M., Wang, Y., Zhang Y., Basche, R.J. and Bassingthwaighte, J.B. (1996) Modelling regional myocardial flows from residue functions of an intravascular indicator. Am J Physiol 271, H1643-H1655.

Fritz-Hansen, T., Rostrup, E., Larsson, H.B., Sondergaard, L., Ring, P. and Hendriksen, O. (1996) Measurement of the arterial concentration Gd-DTPA using MRI; a step toward quantitative perfusion imaging. Magn Reson Med 36, 347-357.