Compartmental Models for Kinetic Parameter Estimation

A selection of parametric models are provided that combine a compartmental model for tissue and a functional form of the arterial input function.

compartmentalModel(type)

Arguments

type
is a character string that identifies the type of compartmental model to be used. Acceptable models include:

"weinmann"

Weinmann AIF convolved with a single compartment (Kety) model

"extended"

Kety model extended with additional vascular compartment (default)

"orton.exp"

Extended model using Orton's exponential arterial input function

"orton.cos"

Extended model using Orton's raised cosine arterial input function

"kety.orton.exp"

Kety model using Orton's exponential arterial input function

"kety.orton.cos"

Kety model using Orton's raised cosine arterial input function

"weinmann.empirical"

User-specified empirical AIF convolved with a single compartment model

"extended.empirical"

Extended model with user-specified empirical arterial input function

Value

A function.

Details

Parametric models from the DCE-MRI literature are provided to the user for kinetic parameter estimation. All models, with the exception of those marked ‘empirical’ incorporate a parametric model for the arterial input function (AIF).

See also