Double-Angle Method for B1+ Mapping

For in vivo MRI at high field (\(\geq3\) T) it is essential to consider the homogeneity of the active B1 field (B1+). The B1+ field is the transverse, circularly polarized component of B1 that is rotating in the same sense as the magnetization. When exciting or manipulating large collections of spins, nonuniformity in B1+ results in nonuniform treatment of spins. This leads to spatially varying image signal and image contrast and to difficulty in image interpretation and image-based quantification.

doubleAngleMethod(low, high, low.deg)

Arguments

low
is the (3D) array of signal intensities at the low flip angle.

high
is the (3D) array of signal intensities at the high flip angle (note, 2*low = high).

low.deg
is the low flip angle (in degrees).

Value

An array, the same dimension as the acquired signal intensities, is returned containing the multiplicative factor associated with the low flip angle acquisition. That is, if no B1+ inhomogeneity was present then the array would only contain ones. Numbers other than one indicate the extent of the inhomogeneity as a function of spatial location.

Details

The proposed method uses an adaptation of the double angle method (DAM). Such methods allow calculation of a flip-angle map, which is an indirect measure of the B1+ field. Two images are acquired: \(I_1\) with prescribed tip \(\alpha_1\) and \(I_2\) with prescribed tip \(\alpha_2=2\alpha_1\). All other signal-affecting sequence parameters are kept constant. For each voxel, the ratio of magnitude images satisfies $$\frac{I_2(r)}{I_1(r)}=\frac{\sin\alpha_2(r)f_2(T_1,\mbox{TR})}{\sin\alpha_1(r)f_1(T_1,\mbox{TR})}$$ where \(r\) represents spatial position and \(alpha_1(r)\) and \(\alpha_2(r)\) are tip angles that vary with the spatially varying B1+ field. If the effects of \(T_1\) and \(T_2\) relaxation can be neglected, then the actual tip angles as a function of spatial position satisfy $$\alpha(r)=\mbox{arccos}\left(\left|\frac{I_2(r)}{2I_1(r)}\right|\right)$$ A long repetition time (\(TR\leq{5T_1}\)) is typically used with the double-angle methods so that there is no \(T_1\) dependence in either \(I_1\) or \(I_2\) (i.e., \(f_1(T_1,TR)=f_2(T_1,TR)=1.0\)). Instead, the proposed method includes a magnetization-reset sequence after each data acquisition with the goal of putting the spin population in the same state regardless of whether the or \(\alpha_2\) excitation was used for the preceding acquisition (i.e., \(f_1(T_1,TR)=f_2(T_1,TR)\ne1.0\)).

References

Cunningham, C.H., Pauly, J.M. and Nayak, K.S. (2006) Saturated Double-Angle Method for Rapid B1+ Mapping, Magnetic Resonance in Medicine, 55, 1326-1333.

Author

Brandon Whitcher bwhitcher@gmail.com