Quantitative T2 Methods

The regional blood volume is found by integrating of the tissue concentration curve and the artieral input funciton (AIF). In order to avoid reperfusion effects on the rCBV measurements, the tissue and arteiral concentration curves must first be reduced to their first-pass versions.

T2.fast(cpmg, ...)

# S4 method for array
T2.fast(
  cpmg,
  cpmg.mask,
  TE,
  control = minpack.lm::nls.lm.control(maxiter = 150),
  multicore = FALSE,
  verbose = FALSE
)

T2.lm(signal, TE, guess, control = minpack.lm::nls.lm.control())

Arguments

cpmg
is a multidimensional array of signal intensities. The last dimension is assumed to be a function of the echo times, while the previous dimenions are assued to be spatial.

...
Additional variables defined by the method.

cpmg.mask
is a (logical) multidimensional array that identifies the voxels to be analyzed.

TE
is the vector of echo times (in seconds).

control
An optional list of control settings for nls.lm. See nls.lm.control for the names of the settable control values and their effect.

multicore
is a logical variable (default = FALSE) that allows parallel processing via multicore.

verbose
is a logical variable (default = FALSE) that allows text-based feedback during execution of the function.

signal
is the vector of signal intensities as a function of echo times.

guess
is the vector of initial values for the parameters of interest: \(\rho\) and \(T2\).

Value

A list structure is produced with (all or some of the) parameter estimates

rho

Scaling factor between signal intensity and T2 (proton density).

T2

T2 relaxation time.

References

Kennan, R.P. and J\"ager, H.R. (2004) $T_2$- and $T_2^*$-w DCE-MRI: Blood Perfusion and Volume Estimation using Bolus Tracking, in Quantiative MRI of the Brain (P. Tofts ed.), Wiley: Chichester, UK, pp. 365-412.

See also

Author

Brandon Whitcher bwhitcher@gmail.com